requirements.txt
aio-pika==9.4.1
aiormq==6.8.0
at-config @ git+https://github.com/grigandal625/AT_CONFIG.git@bcea060e2c92a28cfd2c872b43dddeffb544fabb
at-queue @ git+https://github.com/grigandal625/AT_QUEUE.git@1fbbef9242940ef04ab71ce8dc484bae7e166e12
attrs==23.2.0
contextlib2==21.6.0
idna==3.7
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
multidict==6.0.5
pamqp==3.3.0
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.0
schema==0.7.5
yarl==1.9.4
